Fix non-null terminated string property value printing bug
Device properties may be strings that are not terminated by a null character (many properties on a real Mac are like this). Therefore, they need to be printed up to their length, instead of up to a null character.

This is only a problem when the -s flag is used with the infile outfile conversion action.

If there is a known device property that is a string terminated by a null character, then this will need to be revisited. If they exist, then perhaps they can have a \0 suffix or something similar to represent the trailing null character.