Add guard against badly formed vendor paths
The Mac mini 2018 has a couple device paths that are not of the expected format. This can cause gfxutil to hang. Of course, there are many places in the DevicePathToText library that may misbehave when given a badly formed path, but the point of these fixes is to handle only the cases that actually exist since getting Apple to fix the firmware won't happen and that wouldn't help those that have firmwares that give these paths.

These are the paths:
02010C00D041030A0000000001010600001E030A0500017FFF0400
02010C00D041030A0000000001010600001E030A0500027FFF0400
PciRoot(0x0)/Pci(0x1E,0x0)/VenMsg(,01)
PciRoot(0x0)/Pci(0x1E,0x0)/VenMsg(,02)

The device at 00:1E.0 is a 8086:a328 which is an Intel Cannon Lake PCH Serial IO UART Host Controller.

Changes:

edk2misc.c
- Implement new overrides for DevPathToTextVendor and ConvertFromTextVendor.
DevPathToTextVendor uses the old implementation if the size is large enough for an included UUID. Otherwise it assumes the UUID is missing and outputs something like this: VenMsg(,xx) where xx is the remaining bytes if any.
ConvertFromTextVendor is rewritten to accept an empty UUID for the first parameter.

edk2.overrides
- Enable new overrides for DevPathToTextVendor and ConvertFromTextVendor by renaming old implementations in edk2 with "Old" prefix and adding prototype for Old implementations.

gfxutil.xcscheme
- Added a couple sample arguments for testing this, one for path to text and another for text to path.